large scale run works, this commit prints out debugging information for all rnaks when in GPU driver, main modification was to limit the reads up to 3000

+++ src/gpu_loc_assem/helper.hpp
@@ -59,7 +59,7 @@ struct ReadSeq {
 };
 
 struct CtgWithReads {
-  int32_t cid;
+  int64_t cid;
   std::string seq;
   double depth;
   int max_reads;
